cosmetic changes to tests

+++ v2/tests/CurlTest.php
@@ -16,23 +16,23 @@ class CurlTest extends TestCase
     {
         foreach ($this->urls as $url) {
             // Make sure each site is protected by CF.
-            $ch = curl_init($url);
-            curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-            curl_setopt($ch, CURLOPT_FOLLOWLOCATION, true);
-            curl_setopt($ch, CURLOPT_HEADER, 1);
-            curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, 0);
-            curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, 0);
+            $ch = curl_init( $url );
+            curl_setopt( $ch, CURLOPT_RETURNTRANSFER, true );
+            curl_setopt( $ch, CURLOPT_FOLLOWLOCATION, true );
+            curl_setopt( $ch, CURLOPT_HEADER, 1 );
+            curl_setopt( $ch, CURLOPT_SSL_VERIFYHOST, 0 );
+            curl_setopt( $ch, CURLOPT_SSL_VERIFYPEER, 0 );
 
             // Set proxy server if one is available.
             $proxy = $this->getProxyServer();
 
-            if (isset($proxy))
-                curl_setopt($ch, CURLOPT_PROXY, $proxy);
+            if (isset( $proxy ))
+                curl_setopt( $ch, CURLOPT_PROXY, $proxy );
 
-            curl_setopt($ch, CURLOPT_USERAGENT, $this->getAgent());
-            curl_exec($ch);
+            curl_setopt( $ch, CURLOPT_USERAGENT, $this->getAgent() );
+            curl_exec( $ch );
 
-            $this->assertEquals($url.": "."503", $url.": ".curl_getinfo($ch, CURLINFO_HTTP_CODE));
+            $this->assertEquals( $url.": "."503", $url.": ".curl_getinfo( $ch, CURLINFO_HTTP_CODE ) );
         }
     }
 
@@ -53,42 +53,45 @@ class CurlTest extends TestCase
         foreach ($this->urls as $url) {
 
             // Parse url into components.
-            $url_components = parse_url($url);
+            $url_components = parse_url( $url );
 
             // Get cache file (path included).
-            $cache_file = __DIR__ . '/../var/cache/' . md5($url_components['host']);
-            file_exists($cache_file) && unlink($cache_file);
+            $cache_file = __DIR__ . '/../var/cache/' . md5( $url_components['host'] );
+            
+            if (file_exists( $cache_file ))
+                unlink( $cache_file );
 
             // Bypass each site using CFCurl wrapper.
-            $ch = curl_init($url);
-            curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-            curl_setopt($ch, CURLOPT_FOLLOWLOCATION, true);
-            curl_setopt($ch, CURLOPT_HEADER, 1);
-            curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, 0);
-            curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, 0);
-            curl_setopt($ch, CURLOPT_HTTPHEADER, [
+            $ch = curl_init( $url );
+            curl_setopt( $ch, CURLOPT_RETURNTRANSFER, true );
+            curl_setopt( $ch, CURLOPT_FOLLOWLOCATION, true );
+            curl_setopt( $ch, CURLOPT_HEADER, 1 );
+            curl_setopt( $ch, CURLOPT_SSL_VERIFYHOST, 0 );
+            curl_setopt( $ch, CURLOPT_SSL_VERIFYPEER, 0 );
+            curl_setopt( $ch, CURLOPT_HTTPHEADER, [
                 'Proxy-Connection' => null
-            ]);
+            ] );
             
             // Set a proxy server if one is available.
             $proxy = $this->getProxyServer();
 
             if (isset($proxy)) {
-                curl_setopt($ch, CURLOPT_PROXY, $proxy);
+                curl_setopt( $ch, CURLOPT_PROXY, $proxy );
             }
 
-            curl_setopt($ch, CURLOPT_USERAGENT, $this->getAgent());
+            curl_setopt( $ch, CURLOPT_USERAGENT, $this->getAgent() );
 
-            $response = $wrapper->exec($ch);
+            $response = $wrapper->exec( $ch );
 
-            $this->assertEquals($url.": "."200", $url.": ".curl_getinfo($ch, CURLINFO_HTTP_CODE));
-            $this->assertEquals(true, file_exists($cache_file));
-            $this->assertEquals(true, strpos(file_get_contents($cache_file), "cf_clearance"));
+            $this->assertEquals( $url.": "."200", $url.": ".curl_getinfo( $ch, CURLINFO_HTTP_CODE ) );
+            
+            $this->assertEquals( true, file_exists( $cache_file ) );
+            $this->assertEquals( true, strpos( file_get_contents( $cache_file ), "cf_clearance" ) );
 
             // Remove the file from cache.
-            unlink($cache_file);
+            unlink( $cache_file );
 
-            curl_close($ch);
+            curl_close( $ch );
         }
     }
 
@@ -108,33 +111,33 @@ class CurlTest extends TestCase
         foreach ($this->urls as $url) {
 
             // Bypass each site using CFCurl wrapper.
-            $ch = curl_init($url);
-            curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-            curl_setopt($ch, CURLOPT_FOLLOWLOCATION, true);
-            curl_setopt($ch, CURLOPT_HEADER, 1);
-            curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, 0);
-            curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, 0);
+            $ch = curl_init( $url );
+            curl_setopt( $ch, CURLOPT_RETURNTRANSFER, true );
+            curl_setopt( $ch, CURLOPT_FOLLOWLOCATION, true );
+            curl_setopt( $ch, CURLOPT_HEADER, 1 );
+            curl_setopt( $ch, CURLOPT_SSL_VERIFYHOST, 0 );
+            curl_setopt( $ch, CURLOPT_SSL_VERIFYPEER, 0 );
 
-            $url_components = parse_url($url);
+            $url_components = parse_url( $url );
 
             // Get cache file (path included)
-            $cache_file = __DIR__ . '/../var/cache/' . md5($url_components['host']);
+            $cache_file = __DIR__ . '/../var/cache/' . md5( $url_components['host'] );
 
             // Set a proxy server if one is available.
             $proxy = $this->getProxyServer();
 
-            if (isset($proxy)) {
-                curl_setopt($ch, CURLOPT_PROXY, $proxy);
+            if (isset( $proxy )) {
+                curl_setopt( $ch, CURLOPT_PROXY, $proxy );
             }
 
-            curl_setopt($ch, CURLOPT_USERAGENT, $this->getAgent());
+            curl_setopt( $ch, CURLOPT_USERAGENT, $this->getAgent() );
 
-            $response = $wrapper->exec($ch);
+            $response = $wrapper->exec( $ch );
 
-            $this->assertEquals($url.": "."200", $url.": ".curl_getinfo($ch, CURLINFO_HTTP_CODE));
-            $this->assertEquals(false, file_exists($cache_file));
+            $this->assertEquals( $url.": "."200", $url.": ".curl_getinfo( $ch, CURLINFO_HTTP_CODE ));
+            $this->assertEquals( false, file_exists( $cache_file ) );
 
-            curl_close($ch);
+            curl_close( $ch );
         }
     }
 }
